refactoring data-person.json
