+++ src/app/components/about/about.component.html
@@ -1,4 +1,4 @@
 <div class="container-about">
     <h4>About</h4>
-    <p>{{ about.about }}</p>
+    <p>{{ about }}</p>
     <div>
