+++ src/app/components/address/address.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Address } from 'src/app/models/address.model';
-import * as addressJson from 'src/assets/address.json';
+import addressJson from 'src/assets/data-personal.json';
 @Component({
   selector: 'app-address',
   templateUrl: './address.component.html',
@@ -10,6 +10,6 @@ export class AddressComponent implements OnInit{
 public address!: Address;
 
 ngOnInit(): void {
-  this.address = addressJson as Address;
+  this.address = addressJson['data-personal'].address as Address;
 }
 }
