+++ src/app/components/info/info.component.html
@@ -1,5 +1,5 @@
 <div class="container">
-  <div *ngFor="let itemList of infosList.infosDataList">
+  <div *ngFor="let itemList of infosList">
     <div class="row">
       <div class="col-12 col-lg-2">
         <app-sideway-title [sectionTitle]="itemList.sectionTitle" class="container-sideway-title">
