+++ src/app/components/profile/profile.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Profile } from 'src/app/models/profile.model';
-import * as profileJson from 'src/assets/profile.json';
+import profileJson from 'src/assets/data-personal.json';
 
 @Component({
   selector: 'app-profile',
@@ -11,6 +11,6 @@ export class ProfileComponent implements OnInit{
   public profile!: Profile;
 
   ngOnInit(): void {
-    this.profile = profileJson as Profile;
+    this.profile = profileJson['data-personal'].profile as Profile;
   }
 }
