+++ src/app/components/skills/skills.component.html
@@ -36,7 +36,7 @@
 </div> -->
 <div class="container-skills">
   <h4>Skills</h4>
-  <div *ngFor="let itemList of skillsList.skillsList">
+  <div *ngFor="let itemList of skillsList">
   <div class="row">
     <div class="col-4">
       <h3>{{itemList.skill}}</h3>
