+++ src/app/models/infos-data-list.model.ts
@@ -1,8 +1,3 @@
-
-export interface InfosDataList {
-  infosDataList: InfosData[];
-}
-
 export interface InfosData {
   sectionTitle: string;
   infos: InfosItem[];
@@ -13,4 +8,3 @@ export interface InfosItem {
   subTitle: string;
   text: string;
 }
-
