+++ src/app/models/skills-list.model.ts
@@ -1,9 +1,4 @@
-
-export interface SkillsList {
-    skillsList: Skill[];
-  }
-  
-  export interface Skill {
-    skill: string;
-    expertise: number;
-  }
+export interface Skill {
+  skill: string;
+  expertise: number;
+}
