--- src/assets/address.json
@@ -1,7 +1,0 @@
-{
-    "street": "Rua Deodoro Fonseca",
-    "number": 1399,
-    "district": "Jardim Paraíso",
-    "city": "Solaris",
-    "state": "RJ"
-}
