--- src/assets/contact.json
@@ -1,5 +1,0 @@
-{
-    "qrCode": "https://qrcg-free-editor.qr-code-generator.com/main/assets/images/websiteQRCode_noFrame.png",
-    "email": "willians@smiths.com",
-    "phoneNumber": 99999999999
-}
