--- src/assets/profile.json
@@ -1,5 +1,0 @@
-{
-    "imagem": "https://static.wixstatic.com/media/nsplsh_6536313674333556626567~mv2_d_6000_4000_s_4_2.jpg/v1/fill/w_925,h_616,al_c,q_85,usm_0.66_1.00_0.01,enc_auto/nsplsh_6536313674333556626567~mv2_d_6000_4000_s_4_2.jpg",
-    "name": "Willians Smiths",
-    "occupation": "design gráfico"
-}
