+++ tsconfig.json
@@ -19,6 +19,7 @@
     "target": "ES2022",
     "module": "ES2022",
     "useDefineForClassFields": false,
+    "allowSyntheticDefaultImports": true,
     "resolveJsonModule": true,
     "lib": [
       "ES2022",
